alf text input for generic search input
